More reworking, swap Rocket with axum.
I've changed some of the responsibilities of the framework. There's no
longer a dependency on prometheus. While I don't think prometheus is a
waste, I believe that it's better to instead work with files directly.
This avoids the problem during network-based testing where the process
of collecting results changes the results. Of course, this doesn't work
when trying to collect block I/O metrics, as now you're writing to disk.
For now, HTTP metrics collection won't be supported, but it may make a
comeback in the future.

I've swapped Rocket for axum since I simply prefer axum after having
worked with rust more.

The controller now parses the config and experiment configuraiton tomls
and then creates internal structs that are more ergonomic to work with.
For example, the Runner struct internally contains the address and the
port of the runner, whereas this information is separated between
[[hosts]] and [[runners]] inside the config files.

There's heaps more to do on this project, but I'd like to end up with: a
library that provides all the functionality of running experiments in a
distributed environment, a CLI that uses said library, and a web
dashboard that uses the library as an alternative to the CLI.