Changes to file upload, unlink dependencies
When files are uploaded, I've added an unintelligent check, which simply
looks at whether a file ends with '.sh', and then makes it executable if
it does. I've also added functionality that deletes symlinks from
variation directories prior to collecting results so all that
dependencies aren't present in the results.