Huge overhaul of command execution
Commands are now executed by structs implementing the CommandExecutor
trait. That currently includes the LocalCommandExecutor,
RemoteCommandExecutor and the UnifiedExecutor, which allows for both a
mixture of Remote and Local commands to be run in parallel with one
another.

Now, when exporters are spawned, they are maintained in a background
greenthread, and they are (currently) killed by a global static boolean
variable when an experiment variation comes to it's end.

The next goal of development is to allow users to follow the
stdout/stderr of any setup/teardown/execute/exporter script that they
are interested in during their experiment. This will require the use of
WebSockets (so that huge files aren't exceptionally expensive), and
other background threads. This may be the biggest development milestone
yet.

Cleanup needs to be done in a future commit since many docs have now
been deprecated.