More implementation, big src/run.rs update
In run.rs, I've added a Current struct to the ExperimentRunner, which is
now able to track all the variable fields during an experiment. This
also meant that I could change all functions called from the main loop
from Self:: to self., which has simplified many of the signatures. I've
added a test to ensure that the state of the Current struct resets
correctly at the end of an experiment.

I've implemented a very naive version of the websocket output tailing
functionality. This needs to be improved significantly in the future,
but prototyping is prototyping *shrug*. At the moment, a websocket
listener in python is complaining about a close frame problem. I suspect
that this stems from me cloning the websocket inside of the handler. I
may have to create some structs that can just forward stdout/stderr into
the same websocket instead.