Websocket reading works, although buggy.
Following the stdout/stderr of a process seems to now function properly.
I have tested it on an exporter (sar) running on a remote system, and
the output is received from sar, starting from the beginning of the
process. Currently, all the output from the process is not coming
through, so this needs to be ironed out in the next commits.