Port configs into a new format
Both the ExperimentConfig and the standard Config structs have been
modified into a new format. Many common features of both configs have
been unified, and a testing framework has been created which should
simplify regression testing well into the future. Whilst I'm sure that
more changes will be required in the future internally, the
public-facing format of the Configs is likely to remain backwards
compatible from this point onwards.

The readme has also been updated to reflect the current state of the
configuration file formats.

Finally, A future change would be to unify the testing for both
ExperimentConfig and Config, to make it even easier to add more
tests/functionality to the testing later on. Adding a name/description
to each of the tests would be handy whilst debugging to quickly find the
test case that broke.