Rename "Thing" To ProcessOutputRequest.
In the long, slow effort to get websockets functioning in a way I'm
happy with, I've replaced "Thing" with ProcessOutputRequest, as I think
it's more descriptive. The actual content of the struct may change in
the future.

I've also made a macro for spawning a stdout/stderr threads and
capturing the output, although this will likely be converted into a
function later.